+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Version 0.4.0:
 
-   * Switched to an external htslib
+   * Switched to an external htslib. It is currently compatible with htslib versions 1.4 through 1.9 (the latest one at the time of release).
    * Added the `--perRead` option, as used in https://www.biorxiv.org/content/10.1101/481283v1
    * Removed the `--maxDepth` option, it's not longer required.